Remove init_* NotificationData wrappers
