+++ circuit_maintenance_parser/__init__.py
@@ -1,5 +1,4 @@
 """Circuit-maintenance-parser init."""
-import logging
 from typing import Type, Optional
 
 from .data import NotificationData
@@ -47,8 +46,6 @@ SUPPORTED_PROVIDERS = (
 SUPPORTED_PROVIDER_NAMES = [provider.get_provider_type() for provider in SUPPORTED_PROVIDERS]
 SUPPORTED_ORGANIZER_EMAILS = [provider.get_default_organizer() for provider in SUPPORTED_PROVIDERS]
 
-logger = logging.getLogger(__name__)
-
 
 def init_provider(provider_type=None) -> Optional[GenericProvider]:
     """Returns an instance of the corresponding Notification Provider."""
@@ -60,33 +57,6 @@ def init_provider(provider_type=None) -> Optional[GenericProvider]:
 
     except NonexistentProviderError:
         return None
-
-
-def init_data_raw(data_type: str, data_content: bytes) -> Optional[NotificationData]:
-    """Returns an instance of NotificationData from one combination of data type and content."""
-    try:
-        return NotificationData.init(data_type, data_content)
-    except Exception:  # pylint: disable=broad-except
-        logger.exception("Error found initializing data raw: %s, %s", data_type, data_content)
-    return None
-
-
-def init_data_email(raw_email_bytes: bytes) -> Optional[NotificationData]:
-    """Returns an instance of NotificationData from a raw email content."""
-    try:
-        return NotificationData.init_from_email_bytes(raw_email_bytes)
-    except Exception:  # pylint: disable=broad-except
-        logger.exception("Error found initializing data from email raw bytes: %s", raw_email_bytes)
-    return None
-
-
-def init_data_emailmessage(email_message) -> Optional[NotificationData]:
-    """Returns an instance of NotificationData from an email message."""
-    try:
-        return NotificationData.init_from_emailmessage(email_message)
-    except Exception:  # pylint: disable=broad-except
-        logger.exception("Error found initializing data from email message: %s", email_message)
-    return None
 
 
 def get_provider_class(provider_name: str) -> Type[GenericProvider]:
@@ -121,9 +91,7 @@ def get_provider_class_from_sender(email_sender: str) -> Type[GenericProvider]:
 
 __all__ = [
     "init_provider",
-    "init_data_raw",
-    "init_data_email",
-    "init_data_emailmessage",
+    "NotificationData",
     "get_provider_class",
     "get_provider_class_from_sender",
     "ProviderError",
