+++ tests/unit/test_data.py
@@ -9,13 +9,19 @@ from circuit_maintenance_parser.data import NotificationData
 dir_path = os.path.dirname(os.path.realpath(__file__))
 
 
-def test_init():
-    """Test the simple init class."""
-    data = NotificationData.init("my_type", b"my_content")
+def test_init_from_raw():
+    """Test the init_data_raw function."""
+    data = NotificationData.init_from_raw("my_type", b"my_content")
     assert isinstance(data, NotificationData)
     assert len(data.data_parts) == 1
     assert data.data_parts[0].type == "my_type"
     assert data.data_parts[0].content == b"my_content"
+
+
+def test_init_data_raw_with_issue():
+    """Test the init_data_raw function with issue."""
+    data = NotificationData.init_from_raw({}, {})
+    assert data is None
 
 
 def test_init_from_email_bytes():
@@ -27,6 +33,12 @@ def test_init_from_email_bytes():
     assert len(data.data_parts) == 5
 
 
+def test_init_from_email_with_issue():
+    """Test the init_data_email function with issue."""
+    data = NotificationData.init_from_email_bytes("")
+    assert data is None
+
+
 def test_init_from_emailmessage():
     """Test the emailmessage data load."""
     with open(Path(dir_path, "data", "email", "test_sample_message.eml"), "rb") as email_file:
@@ -36,3 +48,9 @@ def test_init_from_emailmessage():
     data = NotificationData.init_from_emailmessage(email_message)
     assert isinstance(data, NotificationData)
     assert len(data.data_parts) == 5
+
+
+def test_init_from_emailmessage_with_issue():
+    """Test the init_data_emailmessage function with issue."""
+    data = NotificationData.init_from_emailmessage("")
+    assert data is None
